# openapi_spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Platform God API",
    "description": "REST API for agent-driven repository governance",
    "version": "0.0.0"
  },
  "paths": {
...
    "/api/v1/runs": {
      "get": {
        "tags": [
          "Runs"
        ],
        "summary": "List Runs",
        "description": "List chain execution runs.\n\nSupports filtering by repository, chain name, and status.\n\nQuery Parameters:\n    repository_root: Filter by repository root path\n    chain_name: Filter by chain name\n    status: Filter by run status (pending, running, completed, failed, cancelled)\n    limit: Maximum results (default: 50, max: 500)\n    offset: Number of results to skip (default: 0)\n\nReturns:\n    RunListResponse with matching runs",
        "operationId": "list_runs_api_v1_runs_get",
        "parameters": [
          {
            "name": "repository_root",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by repository root",
              "title": "Repository Root"
            },
            "description": "Filter by repository root"
          },
          {
            "name": "chain_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by chain name",
              "title": "Chain Name"
            },
            "description": "Filter by chain name"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by run status",
              "title": "Status"
            },
            "description": "Filter by run status"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Maximum results",
              "default": 50,
              "title": "Limit"
            },
            "description": "Maximum results"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Results to skip",
              "default": 0,
              "title": "Offset"
            },
            "description": "Results to skip"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/registry": {
      "get": {
        "tags": [
          "Registry"
        ],
        "summary": "List Entities",
        "description": "List registry entities.\n\nSupports filtering by entity type.\n\nQuery Parameters:\n    entity_type: Filter by entity type (e.g., 'repository', 'agent', 'finding')\n    limit: Maximum results (default: 100, max: 1000)\n    offset: Number of results to skip (default: 0)\n\nReturns:\n    RegistryListResponse with matching entities",
        "operationId": "list_entities_api_v1_registry_get",
        "parameters": [
          {
            "name": "entity_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by entity type",
              "title": "Entity Type"
            },
            "description": "Filter by entity type"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "description": "Maximum results",
              "default": 100,
              "title": "Limit"
            },
            "description": "Maximum results"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Results to skip",
              "default": 0,
              "title": "Offset"
            },
            "description": "Results to skip"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistryListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Registry"
        ],
        "summary": "Register Entity",
        "description": "Register a new entity in the registry.\n\nArgs:\n    request: Entity registration request\n\nReturns:\n    RegistryOperationResponse with operation result\n\nRaises:\n    ConflictError: If entity already exists",
        "operationId": "register_entity_api_v1_registry_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistryEntityRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistryOperationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RegistryEntityRequest": {
        "properties": {
          "entity_type": {
            "type": "string",
            "title": "Entity Type",
            "description": "Type of entity (e.g., 'repository', 'agent', 'finding')",
            "examples": [
              "repository"
            ]
          },
          "entity_id": {
            "type": "string",
            "title": "Entity Id",
            "description": "Unique identifier for the entity",
            "examples": [
              "repo_abc123"
            ]
          },
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data",
            "description": "Entity data"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata",
            "description": "Optional metadata"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "entity_type",
          "entity_id",
          "data"
        ],
        "title": "RegistryEntityRequest",
        "description": "Request to create or update a registry entity."
      },
      "RegistryEntityResponse": {
        "properties": {
          "entity_id": {
            "type": "string",
            "title": "Entity Id",
            "description": "Entity identifier"
          },
          "entity_type": {
            "type": "string",
            "title": "Entity Type",
            "description": "Entity type"
          },
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data",
            "description": "Entity data"
          },
          "created_at": {
            "type": "string",
            "title": "Created At",
            "description": "Creation timestamp"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At",
            "description": "Last update timestamp"
          },
          "checksum": {
            "type": "string",
            "title": "Checksum",
            "description": "Data integrity checksum"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata",
            "description": "Entity metadata"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "entity_id",
          "entity_type",
          "data",
          "created_at",
          "updated_at",
          "checksum"
        ],
        "title": "RegistryEntityResponse",
        "description": "Response model for a registry entity."
      },
      "RegistryListResponse": {
        "properties": {
          "entities": {
            "items": {
              "$ref": "#/components/schemas/RegistryEntityResponse"
            },
            "type": "array",
            "title": "Entities",
            "description": "List of entities"
          },
          "entity_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entity Type",
            "description": "Entity type filter"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Total count"
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "description": "Response limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset",
            "description": "Response offset"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "entities",
          "total",
          "limit",
          "offset"
        ],
        "title": "RegistryListResponse",
        "description": "Response model for listing registry entities."
      },
      "RegistryOperationResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Operation status"
          },
          "operation": {
            "type": "string",
            "title": "Operation",
            "description": "Operation performed"
          },
          "entity_type": {
            "type": "string",
            "title": "Entity Type",
            "description": "Entity type"
          },
          "entity_id": {
            "type": "string",
            "title": "Entity Id",
            "description": "Entity ID"
          },
          "audit_ref": {
            "type": "string",
            "title": "Audit Ref",
            "description": "Audit log reference"
          },
          "checksum": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checksum",
            "description": "Entity checksum after operation"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error",
            "description": "Error message if failed"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status",
          "operation",
          "entity_type",
          "entity_id",
          "audit_ref"
        ],
        "title": "RegistryOperationResponse",
        "description": "Response model for registry operations."
      },
...
      "RunListResponse": {
        "properties": {
          "runs": {
            "items": {
              "$ref": "#/components/schemas/RunSummary"
            },
            "type": "array",
            "title": "Runs",
            "description": "List of run summaries"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Total count matching filters"
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "description": "Response limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset",
            "description": "Response offset"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "runs",
          "total",
          "limit",
          "offset"
        ],
        "title": "RunListResponse",
        "description": "Response model for listing runs."
      },
      "RunStatus": {
        "type": "string",
        "enum": [
          "pending",
          "running",
          "completed",
          "failed",
          "cancelled"
        ],
        "title": "RunStatus",
        "description": "Status of a chain run."
      },
      "RunSummary": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id",
            "description": "Unique run identifier"
          },
          "chain_name": {
            "type": "string",
            "title": "Chain Name",
            "description": "Chain name"
          },
          "repository": {
            "type": "string",
            "title": "Repository",
            "description": "Repository name"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus",
            "description": "Run status"
          },
          "started_at": {
            "type": "string",
            "title": "Started At",
            "description": "Start timestamp"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At",
            "description": "Completion timestamp"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms",
            "description": "Duration in milliseconds"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "chain_name",
          "repository",
          "status",
          "started_at"
        ],
        "title": "RunSummary",
        "description": "Summary information for a chain run."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}